refactor(orm): 统一字段类型short/byte为INTEGER，消除(short)强制转型噪音
按照docs-for-ai orm-model-design规范，避免使用tinyint/short类型：
- 16个short/SMALLINT字段改为int/INTEGER
- 9个byte/TINYINT字段改为int/INTEGER
- 同步更新手写Java代码和测试文件
- 重新生成所有下游代码(entity/api/meta/view/sql)

Verification: mvn clean package -DskipTests BUILD SUCCESS, mvn test BUILD SUCCESS

+++ app-mall-api/src/main/java/app/mall/api/beans/LitemallAdInputBean.java
@@ -54,16 +54,16 @@
         }
 
 
-        private Byte _position;
+        private Integer _position;
 
     
         @PropMeta(propId=5)
     
-        public Byte getPosition(){
+        public Integer getPosition(){
             return _position;
         }
 
-        public void setPosition(Byte value){
+        public void setPosition(Integer value){
             this._position = value;
         }
