+++ app-mall-api/src/main/java/app/mall/api/beans/LitemallAdOutputBean.java
@@ -69,16 +69,16 @@
         }
 
 
-        private Byte _position;
+        private Integer _position;
 
     
         @PropMeta(propId=5)
     
-        public Byte getPosition(){
+        public Integer getPosition(){
             return _position;
         }
 
-        public void setPosition(Byte value){
+        public void setPosition(Integer value){
             this._position = value;
         }
