+++ app-mall-api/src/main/java/app/mall/api/beans/LitemallCouponUserOutputBean.java
@@ -55,16 +55,16 @@
         }
 
 
-        private Short _status;
+        private Integer _status;
 
     
         @PropMeta(propId=4)
     
-        public Short getStatus(){
+        public Integer getStatus(){
             return _status;
         }
 
-        public void setStatus(Short value){
+        public void setStatus(Integer value){
             this._status = value;
         }
