+++ app-mall-service/src/test/java/app/mall/service/entity/LitemallCartAggregateTest.java
@@ -10,9 +10,9 @@ public class LitemallCartAggregateTest {
     @Test
     public void testIncreaseNumber() {
         LitemallCart cart = new LitemallCart();
-        cart.setNumber((short) 1);
-        cart.increaseNumber((short) 2);
-        assertEquals((short) 3, cart.getNumber());
+        cart.setNumber(1);
+        cart.increaseNumber(2);
+        assertEquals(3, cart.getNumber());
     }
 
     @Test
@@ -20,9 +20,9 @@ public class LitemallCartAggregateTest {
         LitemallCart cart = new LitemallCart();
         cart.setUserId("1");
         cart.setProductId("2");
-        cart.setNumber((short) 1);
+        cart.setNumber(1);
         assertDoesNotThrow(cart::validateForCheckout);
-        cart.setNumber((short) 0);
+        cart.setNumber(0);
         assertThrows(IllegalStateException.class, cart::validateForCheckout);
     }
 }
